backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pyyaml
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
opencv-python-headless
pytesseract
mistralai
python-dotenv
